reparameterize the beta parameter estimation for easier optimization
While experimenting with different pseudobulking methods, we found that the optimization
of the parameters of the Beta distribution for cis_permute would sometimes fail.
Exploring the optimization, we found that reparameterizing as log(beta_shape_1) as a function
of log(true_dof) made this problem much more well behaved (close to linear).